# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Driver shift diary",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/shifts": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "List Shifts",
        "description": "Shifts that started on this local day, each with its summary.",
        "operationId": "list_shifts_api_shifts_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shift"
                  },
                  "title": "Response List Shifts Api Shifts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "shifts"
        ],
        "summary": "Start Shift",
        "operationId": "start_shift_api_shifts_post",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftStartIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/current": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Current Shift",
        "description": "The open shift, or null.",
        "operationId": "current_shift_api_shifts_current_get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Shift"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Current Shift Api Shifts Current Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/{shift_id}": {
      "get": {
        "tags": [
          "shifts"
        ],
        "summary": "Get Shift",
        "operationId": "get_shift_api_shifts__shift_id__get",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "shifts"
        ],
        "summary": "Update Shift",
        "operationId": "update_shift_api_shifts__shift_id__patch",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "shifts"
        ],
        "summary": "Delete Shift",
        "description": "Deletes the shift together with its trips.",
        "operationId": "delete_shift_api_shifts__shift_id__delete",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/{shift_id}/close": {
      "post": {
        "tags": [
          "shifts"
        ],
        "summary": "Close Shift",
        "operationId": "close_shift_api_shifts__shift_id__close_post",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftCloseIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/drivers/{driver_id}/shifts": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Driver Shifts",
        "operationId": "driver_shifts_api_admin_drivers__driver_id__shifts_get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shift"
                  },
                  "title": "Response Driver Shifts Api Admin Drivers  Driver Id  Shifts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/drivers/{driver_id}/shifts/{shift_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Driver Shift",
        "operationId": "driver_shift_api_admin_drivers__driver_id__shifts__shift_id__get",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Update Driver Shift",
        "operationId": "update_driver_shift_api_admin_drivers__driver_id__shifts__shift_id__patch",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShiftPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shift"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Driver Shift",
        "operationId": "delete_driver_shift_api_admin_drivers__driver_id__shifts__shift_id__delete",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Shift Id"
            }
          },
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Driver Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaymentBreakdown": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "amount": {
            "type": "integer",
            "title": "Amount",
            "default": 0
          }
        },
        "type": "object",
        "title": "PaymentBreakdown"
      },
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "cash",
          "card"
        ],
        "title": "PaymentMethod"
      },
...
      "Shift": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "status": {
            "$ref": "#/components/schemas/ShiftStatus"
          },
          "local_day": {
            "type": "string",
            "format": "date",
            "title": "Local Day"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "summary": {
            "$ref": "#/components/schemas/ShiftSummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start",
          "end",
          "status",
          "local_day",
          "note",
          "summary"
        ],
        "title": "Shift"
      },
      "ShiftCloseIn": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ShiftCloseIn"
      },
      "ShiftDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "status": {
            "$ref": "#/components/schemas/ShiftStatus"
          },
          "local_day": {
            "type": "string",
            "format": "date",
            "title": "Local Day"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "summary": {
            "$ref": "#/components/schemas/ShiftSummary"
          },
          "trips": {
            "items": {
              "$ref": "#/components/schemas/Trip"
            },
            "type": "array",
            "title": "Trips"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start",
          "end",
          "status",
          "local_day",
          "note",
          "summary",
          "trips"
        ],
        "title": "ShiftDetail"
      },
      "ShiftPatch": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ShiftPatch",
        "description": "Changes to a shift; omitted fields keep their values. `end: null` reopens it."
      },
      "ShiftStartIn": {
        "properties": {
          "start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "note": {
            "type": "string",
            "maxLength": 500,
            "title": "Note",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ShiftStartIn",
        "description": "Start a shift now (no fields), at a given time, or enter a past shift (start + end)."
      },
      "ShiftStatus": {
        "type": "string",
        "enum": [
          "open",
          "closed"
        ],
        "title": "ShiftStatus"
      },
      "ShiftSummary": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "revenue": {
            "type": "integer",
            "title": "Revenue"
          },
          "commission": {
            "type": "integer",
            "title": "Commission"
          },
          "net": {
            "type": "integer",
            "title": "Net"
          },
          "cash": {
            "$ref": "#/components/schemas/PaymentBreakdown"
          },
          "card": {
            "$ref": "#/components/schemas/PaymentBreakdown"
          },
          "duration_min": {
            "type": "integer",
            "title": "Duration Min"
          },
          "net_per_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Per Hour"
          }
        },
        "type": "object",
        "required": [
          "count",
          "revenue",
          "commission",
          "net",
          "cash",
          "card",
          "duration_min",
          "net_per_hour"
        ],
        "title": "ShiftSummary"
      },
      "Trip": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "shift_id": {
            "type": "integer",
            "title": "Shift Id"
          },
          "start": {
            "type": "string",
            "format": "date-time",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date-time",
            "title": "End"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          },
          "payment": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "commission": {
            "type": "integer",
            "title": "Commission"
          },
          "commission_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commission Pct"
          }
        },
        "type": "object",
        "required": [
          "id",
          "shift_id",
          "start",
          "end",
          "amount",
          "payment",
          "commission"
        ],
        "title": "Trip",
        "description": "A stored trip."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}